Change queue architecture. The queue will no longer spawn a video subprocess and run / shepherd it asynchronously. Instead of shelling out to python, the queue will just run the VideoProcessor class inline without a subprocess. It will run synchronously in the queue.
The benefit of this approach is that previously, it took ~2.7 seconds for the next video to start playing after a video was done in the queue. Now, the next video starts instantly.

The VideoProcessor, rather than the queue, is now responsible for checking to see if the current video should be skipped. It will check to see if the current video should be skipped once per second in the main video playing loop. Timing data shows that each check only takes ~0.35 ms, which is plenty fast.