Use one database connection per thread. The only process that uses threads right now is the ThreadingHTTPServer. Previously we were using one database connection per process, i.e. all of the threads would share the same database connection. I thought this would be safer to avoid deadlocks in sqlite, but this was just a theoretical concern.
In practice, I found that sharing one database connection across threads resulted in errors:

    sqlite3.ProgrammingError: Recursive use of cursors not allowed.

See:
https://stackoverflow.com/questions/26629080/python-and-sqlite3-programmingerror-recursive-use-of-cursors-not-allowed
https://github.com/Sentdex/socialsentiment/issues/12

Perhaps a scenario that led to these issues was:
1) thread 1 issues a SELECT on the cursor
2) thread 2 issues a SELECT on the cursor

If thread 2 issues the SELECT before thread 1 had the chance to fetch the results from the cursor, we may get this error? I did not verify this.