Fix bug skipping videos introduced by numpy vectorization optimizations When skipping videos, we'd run into exceptions when we cleared the screen: https://gist.github.com/dasl-/1cec724ed3c021eace7444dbd85b9bcd
This was because when we set up the LEDs, we added 8 "fake" pixels to work around weird glitches that made the last few pixels yellow or dimmer IIRC. But when we are setting the LED array manually with numpy, we weren't populating those last 8 fake pixels, thus the LED array became shorter than the driver expected.

I removed the 8 fake pixels, and my pifi LEDs still seem to be fine. So maybe this wasn't necessary, or maybe it was only necessary on Ryan's. Let's see when I test it on Ryan's again. If it's necessary, we can add it back in and figure out how to make numpy populate those 8 fake pixels.